honv & cooccurrence with integral histograms

+++ PedestrianDetection/IntegralHistogram2D.cpp
@@ -5,7 +5,7 @@ void IntegralHistogram2D::create(int width, int height, int binSize, std::functi
 	for (int binX = 0; binX < binSize; binX++) {		
 		for (int binY = 0; binY < binSize; binY++)
 		{
-			ihist[binX].push_back(cv::Mat(height, width, CV_32FC1, cv::Scalar(0)));
+			ihist[binX].push_back(cv::Mat(height, width, CV_64FC1, cv::Scalar(0)));
 		}
 	}
 
@@ -21,10 +21,10 @@ void IntegralHistogram2D::create(int width, int height, int binSize, std::functi
 				for (int binY = 0; binY < binSize; binY++)
 				{
 					auto&  cell = ihist[binX][binY];
-					cell.at<float>(j, i) =
-						(i - 1 < 0 ? 0 : cell.at<float>(j, i - 1))
-						+ (j - 1 < 0 ? 0 : cell.at<float>(j - 1, i))
-						- ((i - 1 < 0 || j - 1 < 0) ? 0 : cell.at<float>(j - 1, i - 1));
+					cell.at<double>(j, i) =
+						(i - 1 < 0 ? 0 : cell.at<double>(j, i - 1))
+						+ (j - 1 < 0 ? 0 : cell.at<double>(j - 1, i))
+						- ((i - 1 < 0 || j - 1 < 0) ? 0 : cell.at<double>(j - 1, i - 1));
 				}
 			}
 
@@ -49,12 +49,12 @@ void IntegralHistogram2D::calculateHistogramIntegral(int x, int y, int w, int h,
 			// determine integral histogram values of bin at [i][j]
 
 			auto& cell = ihist[binX][binY];
-			float A = (minx > 0 && miny > 0) ? cell.at<float>(miny, minx) : 0;
-			float B = (miny > 0) ? cell.at<float>(miny, maxx) : 0;
-			float C = (minx > 0) ? cell.at<float>(maxy, minx) : 0;
-			float D = cell.at<float>(maxy, maxx);
+			double A = (minx > 0 && miny > 0) ? cell.at<double>(miny, minx) : 0;
+			double B = (miny > 0) ? cell.at<double>(miny, maxx) : 0;
+			double C = (minx > 0) ? cell.at<double>(maxy, minx) : 0;
+			double D = cell.at<double>(maxy, maxx);
 
-			float value = A + D - C - B;
+			double value = A + D - C - B;
 			hist[binX][binY] = value;
 		}
 	}
